requirements.txt
pyyaml
irc
